requirements.txt
google-cloud-secret-manager==2.16.3
Flask==2.3.3
roboflow==1.1.3
Pillow==10.1.0
requests==2.26.0
scipy==1.10.1
google-cloud-storage==2.13.0
google-cloud-documentai==2.20.2
google-api-core==2.14.0
gdown==4.7.1
keras==2.13.1
matplotlib==3.7.3
numpy==1.24.3
roboflow==1.1.3
tensorflow==2.13.1